+++ Crawler/Crawler.DAL/ApplicationDbContext.cs
@@ -1,4 +1,5 @@
-﻿using DAL.Entities;
+﻿using Crawler.Domain.Entities;
+using DAL.Entities;
 using System;
 using System.Collections.Generic;
 using System.Data.Entity;
