+++ Crawler/Crawler.DAL/packages.config
@@ -1,0 +1,7 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="EntityFramework" version="6.1.3" targetFramework="net452" />
+  <package id="MySql.Data" version="6.9.8" targetFramework="net452" />
+  <package id="MySql.Data.Entities" version="6.8.3.0" targetFramework="net452" />
+  <package id="MySql.Data.Entity" version="6.9.8" targetFramework="net452" />
+</packages>
