+++ Crawler/Crawler.Domain/Interfaces/IRepository.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Crawler.Domain.Interfaces
+{
+    public interface IRepository<T> where T : class
+    {
+        IEnumerable<T> GetAll();
+        T GetById(int id);
+        void Add(T item);
+        void Update(T item);
+        void Delete(T item);
+    }
+}
